+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/PartsPerHundredThousand.java
@@ -1,7 +1,7 @@
 package com.concordium.sdk.responses.transactionstatus;
 
 import com.concordium.sdk.responses.Fraction;
-import com.concordium.sdk.types.UInt64;
+import com.concordium.sdk.types.UInt32;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
@@ -10,15 +10,16 @@ import lombok.ToString;
 @Getter
 @ToString
 public class PartsPerHundredThousand {
-    private final Fraction value;
+    private final UInt32 numerator;
 
     private final int HUNDRED_THOUSAND = 100000;
-    PartsPerHundredThousand(UInt64 value) {
-        this.value = new Fraction(value, UInt64.from(HUNDRED_THOUSAND));
+
+    PartsPerHundredThousand(UInt32 value) {
+        this.numerator = value;
     }
 
-    public static PartsPerHundredThousand from(long x) {
-        return new PartsPerHundredThousand(UInt64.from(x));
+    public static PartsPerHundredThousand from(int x) {
+        return new PartsPerHundredThousand(UInt32.from(x));
     }
 
 
@@ -28,6 +29,15 @@ public class PartsPerHundredThousand {
      * @return the value as a floating point value.
      */
     public double asDouble() {
-        return getValue().asDouble();
+        return (double) this.numerator.getValue() / (double) HUNDRED_THOUSAND;
+    }
+
+
+    /**
+     * Get the bytes representing the {@link PartsPerHundredThousand}
+     * @return the serialized format
+     */
+    public byte[] getBytes() {
+        return this.numerator.getBytes();
     }
 }