+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 # Unreleased changes
+ - Cleanup the API a bit for configure baker transaction.
  - Fix serialization of `ConfigureDelegation` transaction
  - Remove `AccountNonce` in favor of just using the `Nonce` type across the API.
  - Fix a bug in the serialization of `AccountIndex`